+++ tests/testthat/test_bestNormalize.R
@@ -104,3 +104,33 @@ test_that("bestNormalize works with lambert of type h", {
   expect_true(is.null(b$other_transforms$lambert_s))
   expect_true(!is.null(b$other_transforms$lambert_h))
 })
+
+test_that("options work for bestNormalize", {
+  ## Log_x
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(a = 1)), warn = F))
+  expect_equal(b$other_transforms$log_x$a, 1)
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(a = 100)), warn = F))
+  expect_equal(b$other_transforms$log_x$a, 100)
+  
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(a = 1, b = 5)), warn = F))
+  expect_equal(b$other_transforms$log_x$b, 5)
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(a = 100, b = 10)), warn = F))
+  expect_equal(b$other_transforms$log_x$b, 10)
+  
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(eps = 1)), warn = F))
+  expect_equal(b$other_transforms$log_x$eps, 1)
+  expect_silent(b <- bestNormalize(train, tr_opts = list(log_x = list(eps = 100)), warn = F))
+  expect_equal(b$other_transforms$log_x$eps, 100)
+  
+  ## Sqrt_x
+  expect_silent(b <- bestNormalize(train, tr_opts = list(sqrt_x = list(a = 1)), warn = F))
+  expect_equal(b$other_transforms$sqrt_x$a, 1)
+  expect_silent(b <- bestNormalize(train, tr_opts = list(sqrt_x = list(a = 100)), warn = F))
+  expect_equal(b$other_transforms$sqrt_x$a, 100)
+  
+  ## yeo_johnson
+  expect_silent(b <- bestNormalize(train, tr_opts = list(yeojohnson = list(eps = 0.1)), warn = F))
+  expect_equal(b$other_transforms$yeojohnson$eps, .1)
+  expect_silent(b <- bestNormalize(train, tr_opts = list(yeojohnson = list(eps = .01)), warn = F))
+  expect_equal(b$other_transforms$yeojohnson$eps, .01)
+})